Initial config setup for new react-router pages
