+++ client/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 import ReactGA from 'react-ga'
-import { Router } from 'react-router'
+import { Router, browserHistory } from 'react-router'
 import createBrowserHistory from 'history/lib/createBrowserHistory'
 import { Provider } from 'react-redux'
 import * as reducers from 'reducers'
@@ -60,7 +60,7 @@ class Root extends React.Component {
   render() {
     return (
       <Provider store={store}>
-        <Router children={routes} history={history} />
+        <Router children={routes} history={browserHistory}  />
       </Provider>
     )
   }
